Create bubble_algorythm.py
Famous "Bubble Sorting algorithm" in python. By cheking werether the next element in a list is greater than the element that is being analyzed, elements can be sorted out by switching off spaces when this condition is presented. The programm will terminate once the lsit is indeed sorted by claiming the sortedd variable as True and ending up the while loop.